pyproject.toml
[project]
name = "job-reports"
version = "0.1.0"
description = "Scrape Welcome to the Jungle latest job offers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "openai>=2.29.0",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "selenium>=4.41.0",
    "sendgrid>=6.12.5",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "notebook>=7.5.5",
]
